requirements.txt
amqp==5.0.9
ansicon==1.89.0
APScheduler==3.8.1
arabic-reshaper==2.1.3
arrow==1.2.1
asgiref==3.3.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
blessed==1.19.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
confusable-homoglyphs==3.2.0
cssutils==2.3.0
dj-database-url==0.5.0
Django==3.2.3
django-colorfield==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-inlinecss==0.3.0
django-picklefield==3.0.1
dnspython==2.2.1
future==0.18.2
ghostscript==0.7
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.53.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
Jinja2==3.0.2
jinxed==1.1.0
kombu==5.2.3
MarkupSafe==2.0.1
mercadopago==2.0.7
oauth2client==4.1.3
oauthlib==3.2.0
Pillow==8.2.0
prompt-toolkit==3.0.24
protobuf==3.17.3
psycopg2==2.8.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyIsEmail==1.4.0
pynliner==0.8.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyserial==3.5
python-barcode==0.13.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python3-ghostscript==0.5.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyusb==1.2.1
qrcode==7.3.1
reportlab==3.6.3
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.7.2
six==1.16.0
smmap==5.0.0
soupsieve==2.3
sqlparse==0.4.1
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xmltodict==0.12.0